Updates. See commit details.
The Android SDK now uses the new Java client jar that points to the new BaaS endpoint.

Due to changes due to the new BaaS endpoint the Sample applications needed changes as well to point to valid endpoints and BaaS orgs/apps.

The ActivityFeed sample application was modified to use Firebase Cloud Messaging instead of the deprecated GCM.  The Push sample app was modified already to do so.